Move ScalarProperties to their own module
